+++ GarageBard/Components/MainToolbar.swift
@@ -51,6 +51,7 @@ struct MainToolbar<ViewModel: PlayerViewModelProtocol>: View {
                         arrowEdge: .bottom,
                         content: {
                             PopoverMenu {
+                                // Play mode
                                 PopoverMenuItem(action: {
                                     withAnimation(.spring()) {
                                         vm.playMode = .perform
@@ -63,6 +64,7 @@ struct MainToolbar<ViewModel: PlayerViewModelProtocol>: View {
                                         Image(systemName: "checkmark")
                                     }
                                 }
+                                .help("Send keystrokes to the game")
                                 PopoverMenuItem(action: {
                                     withAnimation(.spring()) {
                                         vm.playMode = .listen
@@ -75,9 +77,13 @@ struct MainToolbar<ViewModel: PlayerViewModelProtocol>: View {
                                         Image(systemName: "checkmark")
                                     }
                                 }
+                                .help("Listen to the song with GarageBard's synthesizer")
+
+                                // Loop
                                 Divider()
+                                    .padding(.vertical, space(1))
                                 PopoverMenuItem(action: {
-                                    if vm.loopMode == .off {
+                                    if vm.loopMode != .song {
                                         vm.loopMode = .song
                                     } else {
                                         vm.loopMode = .off
@@ -89,6 +95,36 @@ struct MainToolbar<ViewModel: PlayerViewModelProtocol>: View {
                                         Image(systemName: "checkmark")
                                     }
                                 }
+                                .help("Loop the current song")
+
+                                PopoverMenuItem(action: {
+                                    if vm.loopMode != .session {
+                                        vm.loopMode = .session
+                                    } else {
+                                        vm.loopMode = .off
+                                    }
+                                }) {
+                                    Text("Loop session")
+                                    Spacer()
+                                    if vm.loopMode == .session {
+                                        Image(systemName: "checkmark")
+                                    }
+                                }
+                                .help("Loop the entire session")
+
+                                // Continuous playback
+                                Divider()
+                                    .padding(.vertical, space(1))
+                                PopoverMenuItem(action: {
+                                    vm.continuousPlayback.toggle()
+                                }) {
+                                    Text("Continuous playback")
+                                    Spacer()
+                                    if vm.continuousPlayback {
+                                        Image(systemName: "checkmark")
+                                    }
+                                }
+                                .help("Play the next song after the current one ends")
                             }
                         }
                     )