+++ CHANGELOG.md
@@ -4,7 +4,15 @@
 
 ### Added
 
-- Ability to loop songs
+- Add ability to play all the songs in the session continuously
+- Add ability to loop the currently playing song
+- Add ability to loop entire session
+
+### Changed
+
+- Improve tack name handling
+- Tweak popover menu style
+- Tweak overlay button style
 
 ## 1.1.1